* commit '58d92814fcf72b9f7dfab9700a4624b44967cd85':
  conditional execution
  move int to long
  fix usage of assertions
  set default client id to "ci"

+++ api/src/test/java/com/ripstech/apiconnector2/service/ApplicationServiceTest.java
@@ -28,7 +28,7 @@ class ApplicationServiceTest extends ApiSettings {
 		ApplicationService service = api.applications();
 		Application application = service.post(ApplicationSend.createPost(ApplicationServiceTest.class.getCanonicalName()))
 				                          .orThrow(ApiException::new);
-		int id = application
+		long id = application
 				         .getId();
 		String changedName = ApplicationServiceTest.class.getName() + "changed";
 		assertEquals(changedName,
